metrics/test/main/java/io/pivotal/service/BuildpackServiceTest.java
package io.pivotal.service;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import java.util.Map;

import org.junit.Before;
import org.junit.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import io.pivotal.utils.TestFileReader;

public class BuildpackServiceTest {

	@InjectMocks
	BuildpackService buildpackService;
	
	@Mock
	CloudFoundryClientService clientService;
	
	@Before
	public void init() {
		MockitoAnnotations.initMocks(this);
	}
	
	@Test
	public void testGetTotalBuildpacksCount() {
		String path = BuildpackService.BUILDPACKS_API_URI;
		Map<String, Object> responseMap = new TestFileReader().getFileContents("buildpacks.json").getResponseMap();
		Integer totalCount = Integer.valueOf(responseMap.get("total_results").toString());
		when(clientService.getResponseMap(path)).thenReturn(responseMap);
		when(clientService.getTotalResults(responseMap)).thenReturn(totalCount);
		
		Integer totalBuildpacks = buildpackService.getTotalBuildpacks();
		
		assertEquals(8, totalBuildpacks.intValue());
		
		verify(clientService, times(1)).getResponseMap(path);
		verify(clientService, times(1)).getTotalResults(responseMap);
	}
	
	
}
